Fixes to the resize algorithm
The onResize method was applying the image width / DOM width ratio to
the minimum width and height calculations. When the ratio was greater
than one, the cropper was scaling more than the amount its handler was
dragged.

The onMouseMove method was allowing the cropper rectangle to be resized
larger than its container when the aspectRatio property of the Cropper
component was used.